Implement changes from PR feedback
