--- packages/jsx-to-string/package.json
@@ -1,39 +1,0 @@
-{
-  "name": "@jsx-email/jsx-to-string",
-  "version": "1.0.0",
-  "publishConfig": {
-    "access": "public"
-  },
-  "description": "Renders JSX to HTML strings without React.",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/shellscape/jsx-email.git",
-    "directory": "packages/jsx-to-string"
-  },
-  "main": "dist/index.js",
-  "module": "dist/es/index.js",
-  "exports": {
-    ".": {
-      "import": "./dist/es/index.js",
-      "require": "./dist/index.js"
-    }
-  },
-  "engines": {
-    "node": ">=18.0.0"
-  },
-  "files": [
-    "dist/**"
-  ],
-  "keywords": [
-    "react",
-    "email"
-  ],
-  "dependencies": {},
-  "devDependencies": {
-    "@types/react": "^18.2.26",
-    "react": "^18.2.0"
-  },
-  "types": "./dist/index.d.ts",
-  "sideEffects": false
-}
