--- packages/jsx-to-string/src/index.ts
@@ -1,1 +1,0 @@
-export * from './jsx-to-string';
