+++ packages/render/src/escape-string.ts
@@ -1,3 +1,8 @@
+/**
+ * Note: Parts of this file are derived from [React](https://github.com/facebook/react).
+ * @license MIT
+ */
+
 const matchHtmlRegExp = /["'&<>]/;
 
 const enum ASCII {
@@ -10,6 +15,9 @@ const enum ASCII {
 
 /**
  * Escapes special characters and HTML entities in a given html string.
+ *
+ * This is a slightly modified version of React's
+ * [`escapeHtml`](https://github.com/facebook/react/blob/dddfe688206dafa5646550d351eb9a8e9c53654a/packages/react-dom-bindings/src/server/escapeTextForBrowser.js#L53)
  *
  * @param string HTML string to escape for later insertion
  */
