+++ packages/render/src/index.ts
@@ -1,2 +1,2 @@
 export * from './render';
-export * from './render-async';
+export * from './jsx-to-string';
