+++ packages/tailwind/package.json
@@ -34,7 +34,7 @@
     "react": "^18.2.0"
   },
   "dependencies": {
-    "@jsx-email/jsx-to-string": "workspace:*",
+    "@jsx-email/render": "workspace:*",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "tw-to-css": "0.0.11"
